rearrange home and map container to reuse map container, add usecontext to keep track of mapref and onload/onunmount
