+++ client/src/components/transitfunction.jsx
@@ -48,3 +48,51 @@ export const fetchGTFSData = async ({ setStops, setShapes }) => {
     },
   });
 };
+
+
+export function drawBusRoute(map, shapes, setPolylines) {
+  if (!map) {
+    console.log("Map instance is null or undefined.");
+    return;
+  }
+
+  // Grouping the shapes by their ID's in order to draw the correct route
+  const shapesByRoute = shapes.reduce((acc, shape) => {
+    if (!acc[shape.shape_id]) {
+      acc[shape.shape_id] = [];
+    }
+
+    acc[shape.shape_id].push({ lat: shape.lat, lng: shape.lng });
+    return acc;
+  }, {});
+
+  const polylinesArray = [];
+
+  // Draws each route
+  Object.keys(shapesByRoute).forEach((routeId) => {
+    const routePath = shapesByRoute[routeId];
+
+    const routePolyLine = new google.maps.Polyline({
+      path: routePath,
+      geodesic: true,
+      strokeColor: "red",
+      strokeOpacity: 0.1,
+      strokeWeight: 2,
+    });
+
+    // Adds the polylines to the map
+    routePolyLine.setMap(map);
+    polylinesArray.push(routePolyLine);
+
+    // Adds an info window to the displayed route when clicked
+    const infoWindow = new google.maps.InfoWindow({
+      content: `<strong>${routeId}</strong>`,
+    });
+
+    routePolyLine.addListener("click", function () {
+      infoWindow.setPosition(routePath[0]); // Opens the window at the start of the route
+      infoWindow.open(map);
+    });
+  });
+  setPolylines(polylinesArray);
+}
