+++ client/src/main.jsx
@@ -1,10 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import Router from "./Router.jsx";
+import { MapProvider } from "./context/MapContext.jsx"
 import { LoadScriptNext } from "@react-google-maps/api";
 import "./index.css";
-
-const libraries = ["places"];
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
@@ -13,7 +12,9 @@ ReactDOM.createRoot(document.getElementById("root")).render(
       libraries={["places"]}
       loadingElement={<div>Loading...</div>} // Provide loading element for better UX
     >
-      <Router />
+      <MapProvider>
+        <Router />
+      </MapProvider>
     </LoadScriptNext>
   </React.StrictMode>,
 );
